refactor: move noodles slider to a separate component
